add spo triple info extraction
